changed int to double: fixed 0 division: used infinity

+++ CPP09/ex01/RPN.hpp
@@ -7,10 +7,10 @@ class RPN
 {
 private:
 	const std::string	expression;
-	std::stack<int>		iks;
+	std::stack<double>		iks;
 
 	void	stackAffect(const std::string& substring);
-	int		splitNstack(const std::string input, const std::string delimiter);
+	double		splitNstack(const std::string input, const std::string delimiter);
 public:
 	RPN(const std::string& expression);
 	RPN(const RPN& copy);
@@ -18,5 +18,5 @@ public:
 
 	RPN&	operator= (const RPN& copy);
 
-	int	calculate();
+	double	calculate();
 };
